Refined the handling of group names in multiple populations
Group labels are now directly taken from the session state, so that all
places that use them (i.e. in UI output and also in the classification
itself) are updated dynamically when the text box changes. Previously
the output of `st.text_input` was used for this, which is only the
current value of the text box which only works for uses downstream of
the text element and only for that particular group. This is because
Streamlit apps rerun the entire script upon each change.