feat: Improved image processing (#20)
* Updated CellPhe and pin to cellpose v3

cellpose v4 is far slower for CPUs owing to the switch to a transformer
architecture

* Upload files to the app rather than relying on having local access

Now the app is Dockerized we can no longer rely on having direct access
to the machine's file store, so files need to be explicitly uploaded to
/ downloaded from the app.

* Bundle JDK with Docker image

This is needed for the Trackmate integration

* User can specify minimum number of frames a cell must be in

* Allow user to download processing outputs

This includes:

  - time-series features (mandatory)
  - frame features (optional)
  - ROIs zip (optional)
  - masks (optional)
  - trackmate CSV (optional)

* User can select one of the inbuilt cellpose models

* Allow custom cellpose models + segment progress bar

This requires not using CellPhe's `segment_images` function for 2
reasons:

  - The former uses `cellpose.Cellpose`, which doesn't allow for custom
    models (`pretrained_model` argument). Instead we need to use
    `cellpose.CellposeModel`
  - This allows for adding a progress bar update after each segmentation

* Bundle maven and trackmate dependencies in image

This removes the ~2 minute wait to install the ImageJ and TrackMate Java
packages upon the first time tracking is run.

* Add cyto3 to image

Saves time downloading it and reduces the need to have an internet
connection at runtime

* Added ability to resume image processing using previous outputs

I.e. if masks are provided, the segmentation can be skipped

* Tided up UI with columns and headings

* Use a color palette with 20 colours in

* Added transparency of 0.7 to scatter plots

To aid visibility when having large numbers of points

* Stitch sep + dim reduction plots using matplotlib

Previously these plots were arranged into a 2x2 grid using Streamlit's
column, which meant that each plot was entirely separate.
They are now arranged into a grid using matplotlib subplots, which
allows for the sharing of a legend, so that the 3 dimensionality
reduction methods can share a legend rather than using 1 each

* Added ability to download multiple populations output

* Removed redundant check for FrameID

* Keep CellID, group, filename, and group_index in outputs

* Can download classification results

* Remove CUDA packages

* Install CPU only PyTorch

This saves ~4GB in image size